Fix login redirects and slim middleware
